COB Reworked V0.2
Hi! we started from scratch.

The clean up process was moved to the screener (removal of comments,
whitespaces, etc).

The token class allows us to classify each element in our cleaned file
as either a literal (pi, alpha, etc), a numeric or operation. we can
expand these further in the future.

The lexer class will now break down our cleaned file into tokens to be
evaluated by the parser.

The parser runs the operations on these tokens and produces the result.